+++ src/ScottPlot5/ScottPlot5 Sandbox/Sandbox.Console/Program.cs
@@ -1,8 +1,19 @@
 ﻿using ScottPlot;
 
 Plot plt = new();
-plt.Add.Signal(Generate.Sin());
-plt.Add.Signal(Generate.Cos());
+plt.Add.Signal(Generate.Sin()).LineStyle.HandDrawn = true;
+plt.Add.Signal(Generate.Cos()).LineStyle.HandDrawn = true;
+
+foreach (var axis in plt.Axes.GetAxes())
+{
+    axis.FrameLineStyle.HandDrawn = true;
+}
+
+plt.Axes.DefaultGrid.XAxisStyle.MajorLineStyle.HandDrawn = true;
+plt.Axes.DefaultGrid.XAxisStyle.MinorLineStyle.HandDrawn = true;
+plt.Axes.DefaultGrid.YAxisStyle.MajorLineStyle.HandDrawn = true;
+plt.Axes.DefaultGrid.YAxisStyle.MajorLineStyle.HandDrawn = true;
+
 
 plt.SavePng("test.png", 600, 300).LaunchFile();
 plt.SavePng("test.png", 600, 300).LaunchInBrowser();